+++ Changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
+## 2024-02-26
+- Implementing the lexer
+
 ## 2024-02-22
 - Start with hello sailor.